+++ tests/test_enrich.py
@@ -1,12 +1,15 @@
 import decimal
 import os
 import time
-import pandas as pd
 
+import dash
+import pandas as pd
+import pytest
 from dash.exceptions import PreventUpdate
 from dash_extensions.enrich import Output, Input, State, CallbackBlueprint, html, DashProxy, NoOutputTransform, Trigger, \
     TriggerTransform, MultiplexerTransform, PrefixIdTransform, callback, clientside_callback, DashLogger, LogTransform, \
-    BlockingCallbackTransform, dcc, ServersideOutputTransform, ServersideOutput, ALL, CycleBreakerTransform
+    BlockingCallbackTransform, dcc, ServersideOutputTransform, ServersideOutput, ALL, CycleBreakerTransform, \
+    CycleBreakerInput, DependencyCollection
 
 
 # region Test utils/stubs
@@ -21,10 +24,19 @@ def _get_basic_dash_proxy(**kwargs) -> DashProxy:
     return app
 
 
-def _bind_basic_callback(app):
+def _bind_basic_callback(app, flex=False):
+    if flex:
+        return _bind_basic_callback_flex(app)
+
     @app.callback(Output("log_server", "children"), Input("btn", "n_clicks"))
     def update_log(n_clicks):
         return n_clicks
+
+
+def _bind_basic_callback_flex(app):
+    @app.callback(output=dict(n=Output("log_server", "children")), inputs=dict(n_clicks=Input("btn", "n_clicks")))
+    def update_log(n_clicks):
+        return dict(n=n_clicks)
 
 
 def _bind_basic_clientside_callback(app):
@@ -46,19 +58,44 @@ def _basic_dash_proxy_test(dash_duo, app, element_ids=None, btn_id="btn"):
 
 # endregion
 
+@pytest.mark.parametrize(
+    'tst, flt',
+    [(dict(a=Input("a", "prop"), b=[Input("b", "prop")], c=dict(ca=[Input("ca1", "prop"), Input("ca2", "prop")])),
+      [Input("a", "prop"), Input("b", "prop"), Input("ca1", "prop"), Input("ca2", "prop")]),
+     ([Input("a", "prop"), Input("b", "prop"), Input("c", "prop")],
+      [Input("a", "prop"), Input("b", "prop"), Input("c", "prop")]),
+     ((Input("a", "prop"), Input("b", "prop"), Input("c", "prop")),
+      [Input("a", "prop"), Input("b", "prop"), Input("c", "prop")])
+     ])
+def test_dependency_collection(tst, flt):
+    dc = DependencyCollection(tst)
+    # Test iteration.
+    for i, d in enumerate(dc):
+        assert d == flt[i]
+    # Test access.
+    for i in range(len(dc)):
+        assert dc[i] == flt[i]
+    # Test modification.
+    dc[0] = Input("b", "prop")
+    assert dc[0] == Input("b", "prop")
+    # Test addition.
+    _ = dc.append(Input("new", "prop"))
+    assert dc[-1] == Input("new", "prop")
+
+
 def test_callback_blueprint():
     # Test single element.
     cbp = CallbackBlueprint(State("s", "prop"), Output("o", "prop"), Input("i", "prop"))
-    assert cbp.outputs == [Output("o", "prop")]
-    assert cbp.inputs == [State("s", "prop"), Input("i", "prop")]
+    assert list(cbp.outputs) == [Output("o", "prop")]
+    assert list(cbp.inputs) == [State("s", "prop"), Input("i", "prop")]
     # Test list element.
     cbp = CallbackBlueprint(
         [State("s", "prop"), State("s2", "prop")],
         [Output("o", "prop")],
         [Input("i", "prop")]
     )
-    assert cbp.outputs == [Output("o", "prop")]
-    assert cbp.inputs == [State("s", "prop"), State("s2", "prop"), Input("i", "prop")]
+    assert list(cbp.outputs) == [Output("o", "prop")]
+    assert list(cbp.inputs) == [State("s", "prop"), State("s2", "prop"), Input("i", "prop")]
     # Test mix.
     cbp = CallbackBlueprint(
         [State("s", "prop"), State("s2", "prop")],
@@ -68,8 +105,8 @@ def test_callback_blueprint():
         [Input("i", "prop")],
         Output("o2", "prop")
     )
-    assert cbp.outputs == [Output("o", "prop"), Output("o2", "prop")]
-    assert cbp.inputs == [State("s", "prop"), State("s2", "prop"), Input("i0", "prop"),
+    assert list(cbp.outputs) == [Output("o", "prop"), Output("o2", "prop")]
+    assert list(cbp.inputs) == [State("s", "prop"), State("s2", "prop"), Input("i0", "prop"),
                           State("s3", "prop"), Input("i", "prop")]
     # Test variables.
     my_input = html.Button()
@@ -78,8 +115,8 @@ def test_callback_blueprint():
         Input(my_input, "n_clicks"),
         Output(my_output, "children")
     )
-    assert cbp.outputs == [Output(my_output, "children")]
-    assert cbp.inputs == [Input(my_input, "n_clicks")]
+    assert list(cbp.outputs) == [Output(my_output, "children")]
+    assert list(cbp.inputs) == [Input(my_input, "n_clicks")]
     # Test kwargs.
     cbp = CallbackBlueprint(
         Input(my_input, "n_clicks"),
@@ -87,6 +124,55 @@ def test_callback_blueprint():
         hello="world"
     )
     assert cbp.kwargs == dict(hello="world")
+
+
+def test_flexible_callback_signature():
+    # Test input/output/state as kwargs.
+    cbp = CallbackBlueprint(
+        output=[Output("o", "prop")],
+        inputs=[Input("i", "prop")],
+        hello="world"
+    )
+    assert list(cbp.inputs) == [Input("i", "prop")]
+    assert list(cbp.outputs) == [Output("o", "prop")]
+    assert cbp.kwargs == dict(hello="world")
+    # Test dict grouping.
+    cbp = CallbackBlueprint(
+        output=dict(o=Output("o", "prop"), u=Output("u", "prop")),
+        inputs=dict(i=Input("i", "prop"), s=State("s", "prop")),
+        hello="world"
+    )
+    assert list(cbp.inputs) == [Input("i", "prop"), State("s", "prop")]
+    assert list(cbp.outputs) == [Output("o", "prop"), Output("u", "prop")]
+    assert cbp.kwargs == dict(hello="world")
+    # Test complex dict grouping.
+    cbp = CallbackBlueprint(
+        output=[Output("o", "prop"), Output("u", "prop")],
+        inputs=dict(w=dict(i=Input("i", "prop"), s=State("s", "prop")), z=dict(i=Input("i2", "prop"))),
+        hello="world"
+    )
+    assert list(cbp.inputs) == [Input("i", "prop"), State("s", "prop"), Input("i2", "prop")]
+    assert list(cbp.outputs) == [Output("o", "prop"), Output("u", "prop")]
+    assert cbp.kwargs == dict(hello="world")
+
+
+def test_flexible_callback_signature_in_app(dash_duo):
+    app = DashProxy()
+    app.layout = html.Div([
+        html.Div(id="a", children="a"),
+        html.Div(id="b", children="b"),
+        html.Div(id="log", children=None),
+    ])
+
+    @app.callback(output=dict(x=Output("log", "children")),
+                  inputs=dict(dict(g=dict(a=Input("a", "children"), b=Input("b", "children")))))
+    def update_x(g):
+        return dict(x=f"{g['a']}_{g['b']}")
+
+    dash_duo.start_server(app)
+    time.sleep(0.1)
+    x = dash_duo.find_element("#log")
+    assert x.text == "a_b"
 
 
 def test_dash_proxy(dash_duo):
@@ -97,13 +183,29 @@ def test_dash_proxy(dash_duo):
     _basic_dash_proxy_test(dash_duo, app)
 
 
-def test_no_output_transform(dash_duo):
+def test_dash_output_input_state_compatibility(dash_duo):
+    app = _get_basic_dash_proxy()
+
+    @app.callback(dash.Output("log_server", "children"),
+                  dash.Input("btn", "n_clicks"),
+                  dash.State("btn", "n_clicks"))
+    def update_log(n_clicks, state):
+        return n_clicks
+
+    _basic_dash_proxy_test(dash_duo, app, element_ids=["log_server"])
+
+
+@pytest.mark.parametrize(
+    'args, kwargs',
+    [([Input("btn", "n_clicks")], dict()),
+     ([], dict(inputs=dict(n_clicks=Input("btn", "n_clicks"))))])
+def test_no_output_transform(dash_duo, args, kwargs):
     app = DashProxy()
     app.layout = html.Div([
         html.Button(id="btn"),
     ])
 
-    @app.callback(Input("btn", "n_clicks"))
+    @app.callback(*args, **kwargs)
     def update(n_clicks):
         return n_clicks
 
@@ -155,7 +257,11 @@ def test_trigger_transform(dash_duo):
     assert log.text == "1-1"
 
 
-def test_multiplexer_transform(dash_duo):
+@pytest.mark.parametrize(
+    'args, kwargs',
+    [([Output("log", "children"), Input("right", "n_clicks")], dict()),
+     ([], dict(output=[Output("log", "children")], inputs=dict(n_clicks=Input("right", "n_clicks"))))])
+def test_multiplexer_transform(dash_duo, args, kwargs):
     app = DashProxy(prevent_initial_callbacks=True, transforms=[MultiplexerTransform()])
     app.layout = html.Div([
         html.Button(id="left"),
@@ -164,9 +270,9 @@ def test_multiplexer_transform(dash_duo):
     ])
     app.clientside_callback("function(x){return 'left'}", Output("log", "children"), Input("left", "n_clicks"))
 
-    @app.callback(Output("log", "children"), Input("right", "n_clicks"))
-    def update_right(_):
-        return "right"
+    @app.callback(*args, **kwargs)
+    def update_right(n_clicks):
+        return ["right"]
 
     # Check that the app works.
     dash_duo.start_server(app)
@@ -211,9 +317,10 @@ def test_multiplexer_transform_wildcard(dash_duo):
     assert dash_duo.find_element(cssid(id="1", type="div")).text == "Hello from group 1"
 
 
-def test_prefix_id_transform(dash_duo):
+@pytest.mark.parametrize('flex', [False, True])
+def test_prefix_id_transform(dash_duo, flex):
     app = _get_basic_dash_proxy(transforms=[PrefixIdTransform(prefix="x")])
-    _bind_basic_callback(app)
+    _bind_basic_callback(app, flex)
     _bind_basic_clientside_callback(app)
     # Check that both server and client side callbacks work.
     _basic_dash_proxy_test(dash_duo, app, ["x-log_server", "x-log_client"], "x-btn")
@@ -232,18 +339,23 @@ def test_global_blueprint(dash_duo):
     _basic_dash_proxy_test(dash_duo, app)
 
 
-def test_blocking_callback_transform(dash_duo):
+@pytest.mark.parametrize(
+    'args, kwargs',
+    [([Output("log", "children"), Input("trigger", "n_intervals")], dict()),
+     ([], dict(output=[Output("log", "children")],
+               inputs=dict(tick=Input("trigger", "n_intervals"))))])
+def test_blocking_callback_transform(dash_duo, args, kwargs):
     app = DashProxy(transforms=[BlockingCallbackTransform(timeout=5)])
     app.layout = html.Div([html.Div(id="log"), dcc.Interval(id="trigger", interval=500)])
     msg = "Hello world!"
 
-    @app.callback(Output("log", "children"), Input("trigger", "n_intervals"), blocking=True)
-    def update(_):
+    @app.callback(*args, **kwargs, blocking=True)
+    def update(tick):
         time.sleep(1)
-        return msg
+        return [msg]
 
     # Check that stuff works. It doesn't using a normal Dash object.
-    dash_duo.start_server(app)
+    dash_duo.start_server(app, port=4757)
     dash_duo.wait_for_text_to_equal("#log", msg, timeout=5)
     assert dash_duo.find_element("#log").text == msg
 
@@ -265,7 +377,12 @@ def test_blocking_callback_transform_final_invocation(dash_duo):
     dash_duo.wait_for_text_to_equal("#log", "abc", timeout=5)  # final invocation
 
 
-def test_serverside_output_transform(dash_duo):
+@pytest.mark.parametrize(
+    'args, kwargs',
+    [([ServersideOutput("store", "children"), Input("btn", "n_clicks")], dict()),
+     ([], dict(output=ServersideOutput("store", "children"),
+               inputs=dict(n_clicks=Input("btn", "n_clicks"))))])
+def test_serverside_output_transform(dash_duo, args, kwargs):
     app = DashProxy(prevent_initial_callbacks=True, transforms=[ServersideOutputTransform()])
     app.layout = html.Div([
         html.Button(id="btn"),
@@ -273,8 +390,8 @@ def test_serverside_output_transform(dash_duo):
         html.Div(id="log"),
     ])
 
-    @app.callback(ServersideOutput("store", "children"), Input("btn", "n_clicks"))
-    def update_default(_):
+    @app.callback(*args, **kwargs)
+    def update_default(n_clicks):
         return pd.DataFrame(columns=["A"], data=[1])
 
     @app.callback(Output("log", "children"), Input("store", "children"))
@@ -286,7 +403,7 @@ def test_serverside_output_transform(dash_duo):
     assert dash_duo.find_element("#store").text == ""
     assert dash_duo.find_element("#log").text == ""
     dash_duo.find_element("#btn").click()
-    time.sleep(0.01)  # wait for callback code to execute.
+    time.sleep(0.1)  # wait for callback code to execute.
     assert dash_duo.find_element("#store").text != ""
     assert dash_duo.find_element("#log").text == '{"A":{"0":1}}'
 
@@ -329,10 +446,15 @@ def test_serverside_output_transform_memoize(dash_duo):
     assert dash_duo.find_element("#log2").text == "1"
 
 
-def test_log_transform(dash_duo):
+@pytest.mark.parametrize(
+    'args, kwargs',
+    [([Output("log_server", "children"), Input("btn", "n_clicks")], dict()),
+     ([], dict(output=[Output("log_server", "children")],
+               inputs=dict(n_clicks=Input("btn", "n_clicks"))))])
+def test_log_transform(dash_duo, args, kwargs):
     app = _get_basic_dash_proxy(transforms=[LogTransform(try_use_mantine=False)])
 
-    @app.callback(Output("log_server", "children"), Input("btn", "n_clicks"), log=True)
+    @app.callback(*args, **kwargs, log=True)
     def update_log(n_clicks, dash_logger: DashLogger):
         dash_logger.info("info")
         dash_logger.warning("warning")
@@ -345,7 +467,14 @@ def test_log_transform(dash_duo):
     assert dash_duo.find_element("#log").text == "INFO: info\nWARNING: warning\nERROR: error"
 
 
-def test_cycle_breaker_transform(dash_duo):
+@pytest.mark.parametrize(
+    'c_args, c_kwargs, f_args, f_kwargs',
+    [([Output("celsius", "value"), CycleBreakerInput("fahrenheit", "value")], dict(),
+      [Output("fahrenheit", "value"), Input("celsius", "value")], dict()),
+     ([], dict(output=Output("celsius", "value"), inputs=dict(value=CycleBreakerInput("fahrenheit", "value"))),
+      [], dict(output=Output("fahrenheit", "value"), inputs=dict(value=Input("celsius", "value"))))
+     ])
+def test_cycle_breaker_transform(dash_duo, c_args, c_kwargs, f_args, f_kwargs):
     app = DashProxy(transforms=[CycleBreakerTransform()])
     app.layout = html.Div([
         dcc.Input(id="celsius", type="number"),
@@ -360,11 +489,11 @@ def test_cycle_breaker_transform(dash_duo):
         except ValueError:
             raise PreventUpdate()
 
-    @app.callback(Output("celsius", "value"), Input("fahrenheit", "value", break_cycle=True))
+    @app.callback(*c_args, **c_kwargs)
     def update_celsius(value):
         return str((validate_input(value) - 32) / 9 * 5)
 
-    @app.callback(Output("fahrenheit", "value"), Input("celsius", "value"))
+    @app.callback(*f_args, **f_kwargs)
     def update_fahrenheit(value):
         return str(validate_input(value) / 5 * 9 + 32)